requirements.txt
openai
fastapi
uvicorn
playwright
python-pptx
requests
beautifulsoup4
python-dotenv
